Renamed discord-group to groups
