+++ groups/index.html
@@ -7,9 +7,9 @@
     <link rel="icon" href="/images/index.ico" type="image/x-icon" />
     <title>Discord Groups | Real Dev Squad</title>
     <link rel="stylesheet" href="/global.css" />
-    <link rel="stylesheet" href="/discord-groups/style.css" />
+    <link rel="stylesheet" href="/groups/style.css" />
     <script src="/constants.js"></script>
-    <script type="module" src="/discord-groups/script.js" defer></script>
+    <script type="module" src="/groups/script.js" defer></script>
   </head>
   <body>
     <h3 class="not-verified-tag hidden">
