+++ index.html
@@ -45,9 +45,7 @@
       <a href="/online-members/online-members.html" class="action-button">
         Online Members
       </a>
-      <a class="action-button" href="/discord-groups/index.html">
-        Discord Groups
-      </a>
+      <a class="action-button" href="/groups/index.html"> Discord Groups </a>
       <a class="action-button" href="/standup/index.html"> Standup Updates</a>
     </section>
